Added class for reading in data
This class will contain one method for each of the data sources that we have, based on the differing structures of the data.  The filenames are stored as class constants.
I added the addEdge method to the Graph class to help in reading in the data.  The method takes in two Node IDs and a weight, and creates the Nodes in the Graph if they do not already exist.