Add a private Entry class to Recommender
We faced the problem where we wanted to find the 'items' with the highest cumulative ratings from the users most similar to the provided user. However, no Map or Set would let us sort by values at the end while still keeping the associated nodes (since we would want to get the highest rated items), so we had to implement it in a 'hacky' way: create an Entry class containing a Node and Weight that also implements Comparable so that we can sort on the Entries, while still having a 'map' between Nodes and Weights. We realize that this takes double the space, as we have to store each node twice, but feel that since we are only looking at the item nodes adjacent to the top 10 most similar users to the current user, and because we could think of no other efficient way to solve this problem, that this was a trade-off we were willing to make.

Also make the instance variables in Edge private, as we have getters for those
Also reformatted the code to use spaces instead of tabs
Change the edge weight to be a double, but cast it to be an int in flow (because flow is not polynomial if edge weights are not integers)