Massive refactoring
GraphToolkit.java was getting too big. We want it to just be an API, a repository for different algorithms that can be accessed in one place. We can spin off these algorithms into package-private classes that allow for cleaner code while still giving the outside user access to them through GraphToolkit