Began to add tests for DataReader class
I began to write tests for the DataReader class.  I added in a method in DataReader to read in a sample file that we wrote, with different spacings than the MovieLens DataSet.  I also corrected the method in DataReader that reads in the MovieLens data, since they had movies and users with the same ID, which was messing up our Graph setup.  I also had to make the Edge constructor public so it could be accessed from the tests package, so that I could test the readSampleGraphData method in DataReader.