Added GUI for Recommendation Engine
This GUI enables users to interact directly with the methods in the Recommender class.  The user can select the data set to use (there is currently only one option), the score similarity metric to use, the number of similar users to find, the number of recommendations to return, and the user for which to find recommendations.