Modified score calculation for recommmended items
Our previous metric summed up the values for the Nodes that were adjacent to the X most similar users, where X is 10 by default, and returned those Nodes with the highest score.  However, this is not a fair representation of the rating system, and is thus inaccurate.  For example, if all X users give a movie a rating of 2 out of 5, then that movie will have 2X, a relatively high score, even though it was not really recommended by the similar users.  Thus, we changed the metric to be the total weight of the ratings divided by the square root of the number of users who rated that item.  This metric is more robust, and gives more of a weighted average.  It also favors items that were rated by multiple users, to avoid the problem of a single 5 out of 5 rating being the best recommendation returned.